Calendar Visualization

Calendar Visualization

calendar(
  data = NULL,
  day = NULL,
  value = NULL,
  from = NULL,
  to = NULL,
  ...,
  width = NULL,
  height = NULL,
  elementId = NULL
)

Arguments

data
a data.frame contains day and value for create the calendar.

day, value
column names of day and value. if not provided, it will use the first column as day, and the second column as value.

from
start date

to
end date

...
additional arguments

width, height
Must be a valid CSS unit (like '100%', '400px', 'auto') or a number, which will be coerced to a string and have 'px' appended.

elementId
element id of widgets

Value

a nivo calendar component

See also

Examples

library(nivor)

df <- data.frame(
  day = seq.Date(
    from = as.Date("2020-04-01"),
    length.out = 600,
    by = "days"
  ),
  value = round(runif(600) * 1000, 0)
)

# default
calendar(df)

# GitHub style
calendar(
  data = df,
  emptyColor = "#eeeeee",
  colors = c("#d6e685", "#8cc665", "#44a340", "#1e6823"),
  yearSpacing = 40,
  monthBorderColor = "#ffffff",
  dayBorderWidth = 2,
  dayBorderColor = "#ffffff"
)